@qalshidi commiting for @zghuang

See merge request swmf_software/swmfpy!6

+++ swmfpy/web.py
@@ -13,6 +13,7 @@ import ftplib
 from functools import lru_cache
 import gzip
 from operator import itemgetter
+import os.path
 import shutil
 import urllib
 import urllib.request
@@ -476,18 +477,23 @@ def download_magnetogram_adapt(time, map_type='fixed', **kwargs):
         raise FileNotFoundError('Could not find a file that matches'
                                 + 'the pattern.')
 
+    directory = kwargs.get('download_dir', './')
+    if directory[-1] != '/':
+        directory += '/'
+
     for filename in filenames:
-        # open the file locally
-        directory = kwargs.get('download_dir', './')
-        if directory[-1] != '/':
-            directory += '/'
-        with open(directory + filename, 'wb') as fhandle:
-            # try to download the magnetogram
-            try:
-                ftp.retrbinary('RETR ' + filename, fhandle.write)
-            except ftplib.all_errors:
-                ftp.quit()
-                raise FileNotFoundError('Cannot download ', filename)
+        # Only try to download if the file does not exist
+        if os.path.isfile(directory+filename) == True:
+            raise RuntimeWarning(f'{filename} exists, not downloading')
+        else:
+            # open the file locally
+            with open(directory + filename, 'wb') as fhandle:
+                # try to download the magnetogram
+                try:
+                    ftp.retrbinary('RETR ' + filename, fhandle.write)
+                except ftplib.all_errors:
+                    ftp.quit()
+                    raise FileNotFoundError('Cannot download ', filename)
 
         # unzip the file
         if '.gz' in filename:
